+++ internal/provider/provider.go
@@ -149,6 +149,7 @@ func (p *scalrProvider) Configure(ctx context.Context, req provider.ConfigureReq
 
 func (p *scalrProvider) Resources(_ context.Context) []func() resource.Resource {
 	return []func() resource.Resource{
+		newAgentPoolTokenResource,
 		newTagResource,
 		newWorkspaceResource,
 	}